cleanup default menu
